move all hooks to lib/hooks
