+++ longterm-memory/src/app/memories.tsx
@@ -3,31 +3,6 @@ import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { X } from "lucide-react";
-import { useState, useEffect } from "react";
-
-export function useMemoryStore() {
-  const [memories, setMemories] = useState<Memory[]>([]);
-
-  useEffect(() => {
-    const storedMemories = localStorage.getItem("memories");
-    if (storedMemories) {
-      setMemories(JSON.parse(storedMemories));
-    }
-  }, [setMemories]);
-
-  const deleteMemory = (index: number) => {
-    const newMemories = memories.filter((_, i) => i !== index);
-    localStorage.setItem("memories", JSON.stringify(newMemories));
-    setMemories(newMemories);
-  };
-
-  const updateMemories = (memories: Memory[]) => {
-    localStorage.setItem("memories", JSON.stringify(memories));
-    setMemories(memories);
-  };
-
-  return { memories, deleteMemory, updateMemories };
-}
 
 export interface MemoryProps {
   memories: Memory[];
